+++ client/src/api/Files/File/File.jsx
@@ -26,6 +26,8 @@ const File = () => {
             })
             console.log(response);
 
+
+
             setImage(response.data.file.image)
             setFile(response.data.file);
             setEvent(response.data.file.event);
@@ -36,44 +38,51 @@ const File = () => {
         getFile()
     }, [])
 
-     // *****FUNCION PARA BORRAR*****
-     const deleteFile = async (e) => {
+    // *****FUNCION PARA BORRAR*****
+    const deleteFile = async (e) => {
         e.preventDefault();
-        const response2 = await axios.delete(
-            `http://localhost:5000/api/deleteFile/${fileId}`, {
-            headers: {
-                "Authorization": token
+
+
+        // *****Confirmación*****
+        let option = window.confirm("Seguro que quieres eliminar el archivo???")
+        if (option === true) {
+
+            // *****Hacemos la llamada*****
+            const response2 = await axios.delete(
+                `http://localhost:5000/api/deleteFile/${fileId}`, {
+                headers: {
+                    "Authorization": token
+                }
+            })
+            try {
+
+                setSuccessMessage(response2.data.message)
+
+                setTimeout(() => {
+                    window.location.href = '/'
+                }, 2000)
+
+            } catch (error) {
+                setErrorMessage(error.response2.data.message)
+                setTimeout(() => {
+                    window.location.href = '/file'
+                }, 2000)
             }
-        })
-        try {
-
-            setSuccessMessage(response2.data.message)
-
-            setTimeout(() => {
-                window.location.href = '/'
-            }, 2000)
-
-        } catch (error) {
-            setErrorMessage(error.response2.data.message)
-            setTimeout(() => {
-                window.location.href = '/file'
-            }, 2000)
-        }
+        };
     };
-
     // ******FILES UNLOGGED*****
     const imageUnlogged = () => (
         <div className="file">
             <div className="header">
                 <Header />
             </div>
-            <div className="container">
+            <div className="container centerFile">
                 <div className="fileUnloggTitle text-center"><p>No estás registrado o no estás Logueado</p></div>
                 {/* *****Buttons***** */}
                 <div className="container fileUnloggButtons mb-3">
-                    <div className=" row justify-content-end">
+                    <div className=" row justify-content-start">
                         <div className="col-auto">
-                            <Link className="btn btn-sm btn-primary" type="button" to="/files">Volver</Link>
+                            <Link className="btn btn-primary" type="button" to="/files">Volver</Link>
                         </div>
                     </div>
                 </div>
@@ -87,36 +96,30 @@ const File = () => {
             <div className="header">
                 <Header />
             </div>
-            <div className="container">
-                <div className="fileTitle text-center"><p>IMAGEN</p></div>
+            <div className="container centerFile">
+                <div className="fileTitle text-center"><p>{file.fileName}</p></div>
                 <div className="imgDiv row justify-content-center">
                     <img className="imgFile " src={image.url} alt={file.fileName} />
                 </div>
-                <div className="container fileTable">
-                    <div className="headfile">
-                        <div className="reqfile"><strong>Nombre:</strong></div>
-                        <div className="resfile"> {file.fileName}</div>
-                        <div className="reqfile"><strong>Descripción:</strong></div>
-                        <div className="resfile"> {file.description}</div>
-                        <div className="reqfile"><strong>Fecha</strong></div>
-                        <div className="resfile"> {file.date}</div>
-                        <div className="reqfile"><strong>Evento:</strong></div>
-                        <div className="resfile"> {event.name}</div>
-                        <div className="reqfile"><strong>Usuario:</strong></div>
-                        <div className="resfile"> {user.name} {user.surname}</div>
+                <div className="container fileTable w-100 table table-responsive">
+                    <div className="headFile ">
+                        <div className="reqfile"><strong>Nombre:</strong> {file.fileName}</div>
+                        <div className="reqfile"><strong>Descripción:</strong> {file.description}</div>
+                        <div className="reqfile"><strong>Fecha</strong> {new Date(file.date).toLocaleDateString("es")}</div>
+                        <div className="reqfile"><strong>Evento:</strong> {event.name}</div>
+                        <div className="reqfile"><strong>Usuario:</strong> {user.name} {user.surname}</div>
                     </div>
                 </div>
 
                 {/* *****Buttons***** */}
                 <div className="container fileButtons mb-3">
                     <div className=" row justify-content-between">
-                        <div className="btn-group btn-group-sm col-auto ">
-                            <button className="btn btn-warning" type="submit" >Modificar
-                            </button>
+                        <div className="btn-group col-auto ">
+                            <Link className="btn btn-warning" type="button" key={file._id} to={`/files/updateFile/${fileId}`}>Modificar</Link>
                         </div>
 
                         <div className="col-auto">
-                            <Link className="btn btn-sm btn-primary" type="button" to="/files">Volver</Link>
+                            <Link className="btn btn-primary" type="button" to="/files">Volver</Link>
                         </div>
                     </div>
                 </div>
@@ -131,33 +134,28 @@ const File = () => {
             <div className="header">
                 <Header />
             </div>
-            <div className="container">
+            <div className="container centerFile">
                 <div className="fileTitle text-center"><p>{file.fileName}</p></div>
                 <div className="imgDiv row justify-content-center">
                     <img className="imgFile" src={image.url} alt={image.fileName} />
                 </div>
-                <div className="container fileTable">
+                <div className="container fileTable w-100 table table-responsive">
                     <div className="headfile">
-                        <div className="reqfile"><strong>Nombre:</strong></div>
-                        <div className="resfile"> {file.fileName}</div>
-                        <div className="reqfile"><strong>Descripción:</strong></div>
-                        <div className="resfile"> {file.description}</div>
-                        <div className="reqfile"><strong>Fecha</strong></div>
-                        <div className="resfile"> {file.date}</div>
-                        <div className="reqfile"><strong>Evento:</strong></div>
-                        <div className="resfile"> {event.name}</div>
-                        <div className="reqfile"><strong>Usuario:</strong></div>
-                        <div className="resfile"> {user.name} {user.surname}</div>
+                        <div className="reqfile"><strong>Nombre:</strong> {file.fileName}</div>
+                        <div className="reqfile"><strong>Descripción:</strong> {file.description}</div>
+                        <div className="reqfile"><strong>Fecha</strong> {new Date(file.date).toLocaleDateString("es")}</div>
+                        <div className="reqfile"><strong>Evento:</strong> {event.name}</div>
+                        <div className="reqfile"><strong>Usuario:</strong> {user.name} {user.surname}</div>
                     </div>
                 </div>
 
- {/* *****AVISOS DE ERRORES***** */}
- <div className="message_ok shadow-lg p-1 m-3 bg-body rounded border text-center" style={{ display: successMessage ? "block" : "none" }}>
+                {/* *****AVISOS DE ERRORES***** */}
+                <div className="message_ok shadow-lg p-3 m-3 bg-body rounded border text-center" style={{ display: successMessage ? "block" : "none" }}>
                     <div>
                         {successMessage}
                     </div>
                 </div>
-                <div className="message_nok shadow-lg p-1 m-3 bg-body rounded border text-center" style={{ display: errorMessage ? "block" : "none" }}>
+                <div className="message_ok shadow-lg p-3 m-3 bg-body rounded border text-center" style={{ display: errorMessage ? "block" : "none" }}>
                     <div>
                         {errorMessage}
                     </div>
@@ -166,14 +164,14 @@ const File = () => {
                 {/* *****Buttons***** */}
                 <div className="container fileButtons mb-3">
                     <div className=" row justify-content-between">
-                        <div className="btn-group btn-group-sm col-auto ">
-                            <button className="btn btn-warning" type="submit" >Modificar
-                            </button>
+                        <div className="col-auto">
+                            <Link className="btn btn-primary" type="button" to="/files">Volver</Link>
+                        </div>
+                        <div className="btn-group col-auto ">
+                            <Link className="btn btn-warning" type="button" key={file._id} to={`/files/updateFile/${fileId}`}>Modificar</Link>
                             <button className="btn btn-danger" onClick={deleteFile}>Borrar </button>
                         </div>
-                        <div className="col-auto">
-                            <Link className="btn btn-sm btn-primary" type="button" to="/files">Volver</Link>
-                        </div>
+
                     </div>
                 </div>
             </div>
