+++ client/src/api/Reserves/Reserve/Reserve.jsx
@@ -7,7 +7,7 @@ import axios from "axios";
 
 const Reserve = () => {
 
-    const [reserve_id,setReserve_id] = useState({})
+    const [reserve_id, setReserve_id] = useState({})
     const [event, setEvent] = useState({})
     const [participating, setParticipating] = useState({})
     const token = localStorage.getItem('token')
@@ -16,7 +16,7 @@ const Reserve = () => {
 
     useEffect(() => {
         const getReserve = async () => {
-            const response = await axios.get("http://localhost:5000/api/findReserve", {
+            const response = await axios.get(`http://localhost:5000/api/findReserve/${reserve_id}`, {
                 headers: {
                     "Authorization": token
                 }
@@ -30,56 +30,59 @@ const Reserve = () => {
         getReserve()
     }, [])
 
-// *****FUNCION PARA BORRAR*****
-const deleteReserve = async (e) => {
-    e.preventDefault();
-    const response2 = await axios.delete(
-        `http://localhost:5000/api/deleteReserve/${reserve_id}`, {
-        headers: {
-            "Authorization": token
-        }
-    })
-    try {
+    // *****FUNCION PARA BORRAR*****
+    const deleteReserve = async (e) => {
+        e.preventDefault();
 
-        setSuccessMessage(response2.data.message)
+        // *****Confirmación*****
+        let option = window.confirm("Seguro que quieres eliminar esta Reserva???")
+        if (option === true) {
 
-        setTimeout(() => {
-            window.location.href = '/'
-        }, 2000)
+            // *****Hacemos la llamada*****
 
-    } catch (error) {
-        setErrorMessage(response2.data.error.message)
-        setTimeout(() => {
-            window.location.href = '/user'
-        }, 2000)
-    }
-};
+            const response2 = await axios.delete(
+                `http://localhost:5000/api/deleteReserve/${reserve_id}`, {
+                headers: {
+                    "Authorization": token
+                }
+            })
+            try {
 
+                setSuccessMessage(response2.data.message)
+
+                setTimeout(() => {
+                    window.location.href = '/'
+                }, 2000)
+
+            } catch (error) {
+                setErrorMessage(response2.data.error.message)
+                setTimeout(() => {
+                    window.location.href = '/user'
+                }, 2000)
+            }
+        };
+    };
     return (
         <div className=" reserve">
             <div className="header">
                 <Header />
             </div>
-            <div className="container">
+            <div className="container centerReserve">
                 <div className="reserveTitle text-center mt-3"><p>RESERVA</p></div>
-                <div className="container tablaReserve">
+                <div className="container tablaReserve table table-responsive">
                     <div className="headReserve">
-                        <div ><strong>Evento</strong></div>
-                        <div><strong>Participante</strong></div>
-                    </div>
-                    <div className="bodyReserve">
-                        <div> {event.name}</div>
-                        <div> {participating.name} {participating.surname}</div>
+                        <div ><strong>Evento</strong> {event.name}</div>
+                        <div><strong>Participante</strong> {participating.name} {participating.surname}</div>
                     </div>
                 </div>
 
                 {/* *****AVISOS DE ERRORES***** */}
-                <div className="message_ok shadow-lg p-1 m-3 bg-body rounded border text-center" style={{ display: successMessage ? "block" : "none" }}>
+                <div className="message_ok shadow-lg p-3 m-3 bg-body rounded border text-center" style={{ display: successMessage ? "block" : "none" }}>
                     <div>
                         {successMessage}
                     </div>
                 </div>
-                <div className="message_nok shadow-lg p-1 m-3 bg-body rounded border text-center" style={{ display: errorMessage ? "block" : "none" }}>
+                <div className="message_ok shadow-lg p-3 m-3 bg-body rounded border text-center" style={{ display: errorMessage ? "block" : "none" }}>
                     <div>
                         {errorMessage}
                     </div>
@@ -87,17 +90,18 @@ const deleteReserve = async (e) => {
 
                 {/* *****Buttons***** */}
                 <div className="container reserveButtons mb-3">
+                    <div className="volverReserve">
+                        <Link className="btn btn-sm btn-primary" type="button" to="/reserves">Volver</Link>
+                    </div>
                     <div className="btn-group btn-group-sm col-auto ">
                         <button className="btn btn-warning" type="submit" >Modificar
                         </button>
                         <button className="btn btn-danger" onClick={deleteReserve}>Borrar </button>
-                </div>
-                <div className="volverReserve">
-                    <Link className="btn btn-sm btn-primary" type="button" to="/reserves">Volver</Link>
-                </div>
-            </div>
+                    </div>
 
-        </div>
+                </div>
+
+            </div>
         </div >
     )
 }
